Added test.py and updated director.py
I first added the word attribute making it a private attribute, the displayPlayer attribute which is a list of strings that make up the player, and the play attribute which allows the player to play again.

The idea behind the displayPlayer attribute is that as the user looses lives, the string will print less and less until the lives attribute reaches 0.

I also added a play_again method which prompts the user to play again if their lives equals 0.

I added the test.py file to allow for testing of methods throughout the program since we won't pytest this game.  The file will likely be deleted once the game is complete.